requirements.txt
aiohttp
asyncio
asyncpg
python-dotenv
SQLAlchemy
more-itertools